blank line between syntax, package and import
This will add two blank lines between package and message if there
are no imports. The parser does not provide enough context to only
add a blank line after package if there are imports or to treat all
imports as a single block, or to determine whether an import was the
first one.